requirements.txt
django-cors-headers
dotenv
djangorestframework
djangorestframework-jwt
django
djangorestframework-simplejwt
requests
assemblyai
pyaudio
PyPDF2
langchain_groq
gtts
pygame
langchain
pandas
matplotlib
seaborn
nltk
wordcloud